requirements.txt
cmcrameri==1.9
frechetdist==0.6
lxml==5.3.0
matplotlib==3.10.0
networkx==3.4.2
numpy==2.2.2
nx2ipe==0.0.7
pandas==2.2.3
pdf2image==1.17.0
Pillow==11.1.0
plotnine==0.14.5
Requests==2.32.3
s_gd2==1.8.1
scipy==1.15.1
seaborn==0.13.2
setuptools==75.1.0
wand==0.6.13
tulip-python